revert ConcatTransform to FlattenTransform
